Break an import loop
`serializers` -> `view_utils` -> `mixins` -> `serializers`

The `serializers` -> `view_utils` dependency seemed like the most wrong one
to me (serializers are not views, so they shouldn't depend on view
utilities), so I broke it up by moving some stuff to `utils`.
