+++ cvat/apps/engine/view_utils.py
@@ -4,17 +4,15 @@
 
 # NOTE: importing in the utils.py header leads to circular importing
 
-from typing import Any, Dict, Optional, Type
+from typing import Optional, Type
 
 from django.db.models.query import QuerySet
 from django.http.request import HttpRequest
 from django.http.response import HttpResponse
 from django.shortcuts import get_object_or_404
-from django.utils.http import urlencode
 from rest_framework.decorators import action
 from rest_framework.exceptions import PermissionDenied
 from rest_framework.response import Response
-from rest_framework.reverse import reverse as _reverse
 from rest_framework.serializers import Serializer
 from rest_framework.viewsets import GenericViewSet
 from drf_spectacular.utils import extend_schema
@@ -59,40 +57,6 @@ def make_paginated_response(
     serializer = serializer_type(queryset, **serializer_params)
 
     return response_type(serializer.data)
-
-def reverse(viewname, *, args=None, kwargs=None,
-    query_params: Optional[Dict[str, str]] = None,
-    request: Optional[HttpRequest] = None,
-) -> str:
-    """
-    The same as rest_framework's reverse(), but adds custom query params support.
-    The original request can be passed in the 'request' parameter to
-    return absolute URLs.
-    """
-
-    url = _reverse(viewname, args, kwargs, request)
-
-    if query_params:
-        return f'{url}?{urlencode(query_params)}'
-
-    return url
-
-def build_field_filter_params(field: str, value: Any) -> Dict[str, str]:
-    """
-    Builds a collection filter query params for a single field and value.
-    """
-    return { field: value }
-
-def get_list_view_name(model):
-    # Implemented after
-    # rest_framework/utils/field_mapping.py.get_detail_view_name()
-    """
-    Given a model class, return the view name to use for URL relationships
-    that refer to instances of the model.
-    """
-    return '%(model_name)s-list' % {
-        'model_name': model._meta.object_name.lower()
-    }
 
 def list_action(serializer_class: Type[Serializer], **kwargs):
     params = dict(
